+++ RobotNES/MainWindow.xaml.cs
@@ -1,5 +1,7 @@
 ﻿using NESCore;
 
+using Serilog;
+
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -21,6 +23,7 @@ namespace RobotNES
     public partial class MainWindow : Window
     {
         private readonly NesSystem _nesSystem;
+        private readonly ILogger _logger;
 
         public MainWindow()
         {
@@ -31,17 +34,39 @@ namespace RobotNES
                 File.Delete("RobotNES.log");
             }
 
-            _nesSystem = new NesSystem("nestest.nes");
+            _logger = CreateLogger();
+
+            _nesSystem = new NesSystem("nestest.nes", _logger);
         }
 
         private async void Start_Click(object sender, RoutedEventArgs e)
         {
-            await Task.Run(() => _nesSystem.Run());
+            try
+            {
+                await Task.Run(() => _nesSystem.Run());
+            }
+            catch(Exception ex)
+            {
+                // log any uncatched errors
+                _logger.Fatal(ex, "Unhandled error.");
+            }
         }
 
         private async void Stop_Click(object sender, RoutedEventArgs e)
         {
             await Task.Run(() => _nesSystem.Stop());
+        }
+
+        private static ILogger CreateLogger()
+        {
+            var logger = new LoggerConfiguration()
+                .MinimumLevel.Debug()
+                .WriteTo
+                .Console()
+                .WriteTo.File("RobotNES.log")
+                .CreateLogger();
+
+            return logger;
         }
     }
 }
